pyproject.toml
[tool.poetry]
name = "kit-api"
version = "0.1.1"
description = "An api dedicated to kits."
authors = ["Jonxslays"]
license = "BSD-3-Clause"
include = ["kitapi"]

[tool.poetry.dependencies]
python = "^3.10.0"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
python-dotenv = "^0.19.0"
asyncpg = "^0.24.0"
tortoise-orm = "^0.17.8"
uvloop = "^0.16.0"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
isort = "^5.9.3"
pyright = "^0.0.10"

[tool.black]
line-length = 99

[tool.pyright]
include = ["kitapi"]
typeCheckingMode = "strict"

[build-system]
requires = ["poetry-core>=1.1.11"]
build-backend = "poetry.core.masonry.api"
